Firedrake backend: Move assembly code
