+++ tlm_adjoint/firedrake/__init__.py
@@ -8,6 +8,7 @@ del adjoint, alias, cached_hessian, caches, checkpointing, \
 from .backend import backend, backend_RealType, backend_ScalarType  # noqa: E402,E501,F401
 from .backend_code_generator_interface import linear_solver  # noqa: E402,F401
 from .backend_interface import *    # noqa: E402,F401
+from .assembly import *             # noqa: E402,F401
 from .assignment import *           # noqa: E402,F401
 from .block_system import \
     (ConstantNullspace, DirichletBCNullspace, NoneNullspace, UnityNullspace)  # noqa: E402,E501,F401
@@ -22,6 +23,7 @@ from .projection import *           # noqa: E402,F401
 
 from .backend_patches import *      # noqa: E402,F401
 
-del backend, backend_code_generator_interface, backend_interface, assignment, \
-    block_system, caches, equations, expr, functions, hessian_system, \
-    interpolation, parameters, projection, backend_patches  # noqa: F821
+del backend, backend_code_generator_interface, backend_interface, assembly, \
+    assignment, block_system, caches, equations, expr, functions, \
+    hessian_system, interpolation, parameters, projection, \
+    backend_patches  # noqa: F821
